large class refactored
