+++ Designite_Output/DesignSmells.csv
@@ -3,8 +3,9 @@ Super-Mario-Java-2D-Game,com.TETOSOFT.graphics,ScreenManager,Unutilized Abstract
 Super-Mario-Java-2D-Game,com.TETOSOFT.input,GameAction,Unutilized Abstraction,The tool detected the smell in this class because this class is potentially unused. (Please ignore the smell if the reported class is auto-generated and/or used to serve a specific known purpose.)
 Super-Mario-Java-2D-Game,com.TETOSOFT.input,GameAction,Deficient Encapsulation,The tool detected the smell in this class because the class exposes fields belonging to it with public accessibility. Following fields are declared with public accessiblity: NORMAL; DETECT_INITAL_PRESS_ONLY; STATE_RELEASED; STATE_PRESSED; STATE_WAITING_FOR_RELEASE; name; behavior; amount; state
 Super-Mario-Java-2D-Game,com.TETOSOFT.input,InputManager,Unutilized Abstraction,The tool detected the smell in this class because this class is potentially unused. (Please ignore the smell if the reported class is auto-generated and/or used to serve a specific known purpose.)
-Super-Mario-Java-2D-Game,com.TETOSOFT.input,InputManager,Deficient Encapsulation,The tool detected the smell in this class because the class exposes fields belonging to it with public accessibility. Following fields are declared with public accessiblity: INVISIBLE_CURSOR; MOUSE_MOVE_LEFT; MOUSE_MOVE_RIGHT; MOUSE_MOVE_UP; MOUSE_MOVE_DOWN; MOUSE_WHEEL_UP; MOUSE_WHEEL_DOWN; MOUSE_BUTTON_1; MOUSE_BUTTON_2; MOUSE_BUTTON_3; NUM_MOUSE_CODES; NUM_KEY_CODES; keyActions; mouseActions; mouseLocation; centerLocation; comp; robot; isRecentering
-Super-Mario-Java-2D-Game,com.TETOSOFT.input,InputManager,Insufficient Modularization,The tool detected the smell in this class becuase the class has bloated interface (large number of public methods). Total public methods in the class: 25 public methods
+Super-Mario-Java-2D-Game,com.TETOSOFT.input,InputManager,Deficient Encapsulation,The tool detected the smell in this class because the class exposes fields belonging to it with public accessibility. Following fields are declared with public accessiblity: INVISIBLE_CURSOR; NUM_MOUSE_CODES; NUM_KEY_CODES; keyActions; mouseActions; mouse; centerLocation; comp; robot; isRecentering
+Super-Mario-Java-2D-Game,com.TETOSOFT.input,Mouse,Unutilized Abstraction,The tool detected the smell in this class because this class is potentially unused. (Please ignore the smell if the reported class is auto-generated and/or used to serve a specific known purpose.)
+Super-Mario-Java-2D-Game,com.TETOSOFT.input,Mouse,Deficient Encapsulation,The tool detected the smell in this class because the class exposes fields belonging to it with public accessibility. Following fields are declared with public accessiblity: MOUSE_MOVE_LEFT; MOUSE_MOVE_RIGHT; MOUSE_MOVE_UP; MOUSE_MOVE_DOWN; MOUSE_WHEEL_UP; MOUSE_WHEEL_DOWN; MOUSE_BUTTON_1; MOUSE_BUTTON_2; MOUSE_BUTTON_3; mouseLocation
 Super-Mario-Java-2D-Game,com.TETOSOFT.test,GameCore,Unutilized Abstraction,The tool detected the smell in this class because this class is potentially unused. (Please ignore the smell if the reported class is auto-generated and/or used to serve a specific known purpose.)
 Super-Mario-Java-2D-Game,com.TETOSOFT.tilegame,GameEngine,Unutilized Abstraction,The tool detected the smell in this class because this class is potentially unused. (Please ignore the smell if the reported class is auto-generated and/or used to serve a specific known purpose.)
 Super-Mario-Java-2D-Game,com.TETOSOFT.tilegame,MapLoader,Unutilized Abstraction,The tool detected the smell in this class because this class is potentially unused. (Please ignore the smell if the reported class is auto-generated and/or used to serve a specific known purpose.)
