+++ Designite_Output/MethodMetrics.csv
@@ -54,16 +54,8 @@ Super-Mario-Java-2D-Game,com.TETOSOFT.input,InputManager,setRelativeMouseMode,17
 Super-Mario-Java-2D-Game,com.TETOSOFT.input,InputManager,isRelativeMouseMode,3,1,0
 Super-Mario-Java-2D-Game,com.TETOSOFT.input,InputManager,mapToKey,3,1,2
 Super-Mario-Java-2D-Game,com.TETOSOFT.input,InputManager,mapToMouse,3,1,2
-Super-Mario-Java-2D-Game,com.TETOSOFT.input,InputManager,clearMap,13,5,1
-Super-Mario-Java-2D-Game,com.TETOSOFT.input,InputManager,getMaps,14,5,1
-Super-Mario-Java-2D-Game,com.TETOSOFT.input,InputManager,resetAllGameActions,12,5,0
-Super-Mario-Java-2D-Game,com.TETOSOFT.input,InputManager,getKeyName,3,1,1
-Super-Mario-Java-2D-Game,com.TETOSOFT.input,InputManager,getMouseName,24,10,1
-Super-Mario-Java-2D-Game,com.TETOSOFT.input,InputManager,getMouseX,3,1,0
-Super-Mario-Java-2D-Game,com.TETOSOFT.input,InputManager,getMouseY,3,1,0
 Super-Mario-Java-2D-Game,com.TETOSOFT.input,InputManager,recenterMouse,9,2,0
 Super-Mario-Java-2D-Game,com.TETOSOFT.input,InputManager,getKeyAction,9,2,1
-Super-Mario-Java-2D-Game,com.TETOSOFT.input,InputManager,getMouseButtonCode,12,4,1
 Super-Mario-Java-2D-Game,com.TETOSOFT.input,InputManager,getMouseButtonAction,9,2,1
 Super-Mario-Java-2D-Game,com.TETOSOFT.input,InputManager,keyPressed,7,2,1
 Super-Mario-Java-2D-Game,com.TETOSOFT.input,InputManager,keyReleased,7,2,1
@@ -77,6 +69,14 @@ Super-Mario-Java-2D-Game,com.TETOSOFT.input,InputManager,mouseDragged,3,1,1
 Super-Mario-Java-2D-Game,com.TETOSOFT.input,InputManager,mouseMoved,16,3,1
 Super-Mario-Java-2D-Game,com.TETOSOFT.input,InputManager,mouseWheelMoved,3,1,1
 Super-Mario-Java-2D-Game,com.TETOSOFT.input,InputManager,mouseHelper,13,3,3
+Super-Mario-Java-2D-Game,com.TETOSOFT.input,Mouse,Mouse,3,1,0
+Super-Mario-Java-2D-Game,com.TETOSOFT.input,Mouse,getMouseLocation,3,1,0
+Super-Mario-Java-2D-Game,com.TETOSOFT.input,Mouse,getMouseLocationX,3,1,0
+Super-Mario-Java-2D-Game,com.TETOSOFT.input,Mouse,getMouseLocationY,3,1,0
+Super-Mario-Java-2D-Game,com.TETOSOFT.input,Mouse,setMouseLocationX,3,1,1
+Super-Mario-Java-2D-Game,com.TETOSOFT.input,Mouse,setMouseLocationY,3,1,1
+Super-Mario-Java-2D-Game,com.TETOSOFT.input,Mouse,getMouseName,24,10,1
+Super-Mario-Java-2D-Game,com.TETOSOFT.input,Mouse,getMouseButtonCode,12,4,1
 Super-Mario-Java-2D-Game,com.TETOSOFT.test,GameCore,stop,3,1,0
 Super-Mario-Java-2D-Game,com.TETOSOFT.test,GameCore,run,10,1,0
 Super-Mario-Java-2D-Game,com.TETOSOFT.test,GameCore,lazilyExit,15,1,0
