+++ Designite_Output/TypeMetrics.csv
@@ -4,7 +4,8 @@ Super-Mario-Java-2D-Game,com.TETOSOFT.graphics,AnimFrame,2,0,1,1,8,1,0,0,0.0,1,0
 Super-Mario-Java-2D-Game,com.TETOSOFT.graphics,ScreenManager,1,0,15,14,128,30,0,0,0.13333333333333333,0,0,C:\Users\Dana Myers\Desktop\SE433 Software Quality\Project\Super-Mario-Java-2D-Game\src\com\TETOSOFT\graphics\ScreenManager.java
 Super-Mario-Java-2D-Game,com.TETOSOFT.graphics,Sprite,5,0,14,14,51,14,2,0,0.0,0,0,C:\Users\Dana Myers\Desktop\SE433 Software Quality\Project\Super-Mario-Java-2D-Game\src\com\TETOSOFT\graphics\Sprite.java
 Super-Mario-Java-2D-Game,com.TETOSOFT.input,GameAction,9,2,10,10,58,14,0,0,0.0,0,0,C:\Users\Dana Myers\Desktop\SE433 Software Quality\Project\Super-Mario-Java-2D-Game\src\com\TETOSOFT\input\GameAction.java
-Super-Mario-Java-2D-Game,com.TETOSOFT.input,InputManager,19,10,29,25,243,66,0,0,0.41379310344827586,0,0,C:\Users\Dana Myers\Desktop\SE433 Software Quality\Project\Super-Mario-Java-2D-Game\src\com\TETOSOFT\input\InputManager.java
+Super-Mario-Java-2D-Game,com.TETOSOFT.input,InputManager,10,1,21,17,150,34,0,0,0.5238095238095238,0,0,C:\Users\Dana Myers\Desktop\SE433 Software Quality\Project\Super-Mario-Java-2D-Game\src\com\TETOSOFT\input\InputManager.java
+Super-Mario-Java-2D-Game,com.TETOSOFT.input,Mouse,10,9,8,8,66,20,0,0,0.25,0,0,C:\Users\Dana Myers\Desktop\SE433 Software Quality\Project\Super-Mario-Java-2D-Game\src\com\TETOSOFT\input\Mouse.java
 Super-Mario-Java-2D-Game,com.TETOSOFT.test,GameCore,5,0,9,9,88,10,0,0,0.5555555555555556,0,0,C:\Users\Dana Myers\Desktop\SE433 Software Quality\Project\Super-Mario-Java-2D-Game\src\com\TETOSOFT\test\GameCore.java
 Super-Mario-Java-2D-Game,com.TETOSOFT.tilegame,GameEngine,12,0,14,11,261,49,0,0,0.35714285714285715,0,1,C:\Users\Dana Myers\Desktop\SE433 Software Quality\Project\Super-Mario-Java-2D-Game\src\com\TETOSOFT\tilegame\GameEngine.java
 Super-Mario-Java-2D-Game,com.TETOSOFT.tilegame,MapLoader,9,0,15,8,201,33,0,0,0.6666666666666666,0,1,C:\Users\Dana Myers\Desktop\SE433 Software Quality\Project\Super-Mario-Java-2D-Game\src\com\TETOSOFT\tilegame\MapLoader.java
